requirements.txt
Django==1.8.6
django-extensions==1.5.9
django-filter==0.11.0
djangorestframework==3.3.1
djoser==0.4.0
sh==1.11
six==1.10.0
wheel==0.24.0
